refactored file structure so server content is maintained in server folder with bin folder used to run server files and client folder used for front end client facing work
