+++ Team Development/package.json
@@ -13,13 +13,14 @@
   "author": "Trace Larson, Daniel Hughes, Chris Feidor",
   "license": "MIT",
   "dependencies": {
-    "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
     "concurrently": "^4.0.1",
     "dotenv": "^6.0.0",
+    "nodemon": "^1.18.4",
+    "axios": "^0.18.0",
+    "body-parser": "^1.18.3",
+    "create-react-app": "^2.0.3",
     "express": "^4.16.3",
     "mongoose": "^5.3.1",
-    "nodemon": "^1.18.4",
     "validator": "^10.8.0"
   }
 }
