+++ Team Development/server/app.js
@@ -2,7 +2,6 @@ const express = require('express')
 const app = express()
 const mongoose = require('mongoose')
 const bodyParser = require('body-parser')
-const path = require('path')
 
 
 // Configure Body Parser
